+++ src/main/java/frc/robot/commands/orchestrator/SetPersistentAction.java
@@ -5,11 +5,11 @@
 package frc.robot.commands.orchestrator;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
-import frc.robot.subsystems.OrchestratorV2;
-import frc.robot.subsystems.OrchestratorV2.OrchestratorAction;
+import frc.robot.subsystems.Orchestrator;
+import frc.robot.subsystems.Orchestrator.OrchestratorAction;
 
 public class SetPersistentAction extends InstantCommand {
-  private final OrchestratorV2 m_Orchestrator;
+  private final Orchestrator m_Orchestrator;
   private final OrchestratorAction m_DesiredAction;
 
   /**
@@ -19,11 +19,11 @@ public class SetPersistentAction extends InstantCommand {
    * @param orchestrator The orchestrator subsystem.
    * @param action The action you want the orchestrator to perform.
    */
-  public SetPersistentAction(OrchestratorV2 orchestrator, OrchestratorAction desiredAction)
+  public SetPersistentAction(Orchestrator orchestrator, OrchestratorAction desiredAction)
   {
     this.m_Orchestrator = orchestrator;
     this.m_DesiredAction = desiredAction;
-    
+
     // add requirements to make it so that multiple commands cannot be requested at once
     addRequirements(m_Orchestrator);
   }
